pyproject.toml
[tool.poetry]
name = "rag-app"
version = "0.1.0"
description = "RAG App"
authors = ["Mohan Sharma <email@dummy.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
jinja2 = "^3.1.2"
uvicorn = "^0.30.1"
python-dotenv = "^1.0.1"
gunicorn = "^22.0.0"
dependency-injector = "^4.41.0"
mypy="^1.10.0"
beautifulsoup4 = "^4.12.3"
langchain-community = "^0.2.1"
langchainhub = "^0.1.17"
langchain = "^0.2.1"
langgraph = "^0.0.57"
langchain-openai = "^0.1.8"
langchain-text-splitters = "^0.2.0"
tiktoken = "^0.7.0"
chromadb = "^0.5.0"
tavily-python = "^0.3.3"
black = "^24.4.2"
isort = "^5.13.2"
pydantic = "^2.7.3"
pydantic-settings = "^2.3.1"
pypdf = "^4.2.0"

[tool.poetry.dev-dependencies]
pytest = "^8.2.1"

[build-system]
requires = ["poetry-core>=1.8.3"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "python com/crack/snap/make/app.py"
